LayoutTests/http/tests/websocket/tests/hybi/invalid-subprotocols-expected.txt

CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket 'a b'
CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket 'a,b'
CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket 'a\u0000b'
CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket '\u3042'
CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket 'c d'
CONSOLE MESSAGE: line 1: WebSocket protocols contain duplicates: 'a'
WebSocket should fail if invalid subprotocols are specified

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket(url, 'a b') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket(url, 'a,b') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket(url, 'a\0b') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket(url, '\u3042') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket(url, ['a', 'b', 'c d']) threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket(url, ['a', 'b', 'c', 'a']) threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS successfullyParsed is true

TEST COMPLETE
